+++ src/BitMono.Obfuscation/Abstractions/FileDataWriter.cs
@@ -1,6 +1,6 @@
 ﻿namespace BitMono.Obfuscation.Abstractions;
 
-internal class FileDataWriter : IDataWriter
+public class FileDataWriter : IDataWriter
 {
     public Task WriteAsync(string outputFile, byte[] outputBuffer)
     {