+++ src/BitMono.Protections/DotNetHook.cs
@@ -1,5 +1,6 @@
 ﻿namespace BitMono.Protections;
 
+[UsedImplicitly]
 public class DotNetHook : Protection
 {
     private readonly Renamer _renamer;